# Improve endpoints formatter