--- a/gadgets/trace_tcpretrans/program.bpf.c
+++ b/gadgets/trace_tcpretrans/program.bpf.c
@@ -98,33 +98,33 @@ static __always_inline int __trace_tcp_retrans(void *ctx, const struct sock *sk,
 	family = BPF_CORE_READ(sk, __sk_common.skc_family);
 	switch (family) {
 	case AF_INET:
-		event->src.l3.version = event->dst.l3.version = 4;
+		event->src.version = event->dst.version = 4;
 
-		BPF_CORE_READ_INTO(&event->src.l3.addr.v4, sk,
+		BPF_CORE_READ_INTO(&event->src.addr_raw.v4, sk,
 				   __sk_common.skc_rcv_saddr);
-		if (event->src.l3.addr.v4 == 0)
+		if (event->src.addr_raw.v4 == 0)
 			goto cleanup;
 
-		BPF_CORE_READ_INTO(&event->dst.l3.addr.v4, sk,
+		BPF_CORE_READ_INTO(&event->dst.addr_raw.v4, sk,
 				   __sk_common.skc_daddr);
-		if (event->dst.l3.addr.v4 == 0)
+		if (event->dst.addr_raw.v4 == 0)
 			goto cleanup;
 		break;
 
 	case AF_INET6:
-		event->src.l3.version = event->dst.l3.version = 6;
+		event->src.version = event->dst.version = 6;
 
 		BPF_CORE_READ_INTO(
-			&event->src.l3.addr.v6, sk,
+			&event->src.addr_raw.v6, sk,
 			__sk_common.skc_v6_rcv_saddr.in6_u.u6_addr32);
-		if (((u64 *)event->src.l3.addr.v6)[0] == 0 &&
-		    ((u64 *)event->src.l3.addr.v6)[1] == 0)
+		if (((u64 *)event->src.addr_raw.v6)[0] == 0 &&
+		    ((u64 *)event->src.addr_raw.v6)[1] == 0)
 			goto cleanup;
 
-		BPF_CORE_READ_INTO(&event->dst.l3.addr.v6, sk,
+		BPF_CORE_READ_INTO(&event->dst.addr_raw.v6, sk,
 				   __sk_common.skc_v6_daddr.in6_u.u6_addr32);
-		if (((u64 *)event->dst.l3.addr.v6)[0] == 0 &&
-		    ((u64 *)event->dst.l3.addr.v6)[1] == 0)
+		if (((u64 *)event->dst.addr_raw.v6)[0] == 0 &&
+		    ((u64 *)event->dst.addr_raw.v6)[1] == 0)
 			goto cleanup;
 		break;