This PR improves the way endpoints are formatted in the columns and json formats:

### Testing 

#### Before 

The columns output printed a lot of fields related to the endpoints:

```bash 
$ sudo -E ig run trace_tcpconnect:latest --verify-image=false
INFO[0000] Experimental features enabled
RUNTIME.CONTAINERNA… SRC… SRC… DST… DS… TASK       PID        UID        GID        MNTNS_… TIMESTAMP              SRC.ADDRE… DST.ADDRE…
netem                341… 6    80   6   wget       137202     0          0          4026534 2024-06-25T15:42:05.24 172.17.0.2 1.1.1.1:80
netem                454… 6    443  6   wget       137202     0          0          4026534 2024-06-25T15:42:05.36 172.17.0.2 1.1.1.1:44
netem                606… 6    443  6   wget       137202     0          0          4026534 2024-06-25T15:42:06.44 172.17.0.2 1.0.0.1:4
```

It wasn't possible to choose an endpoint with `--fields`:

```bash
$ sudo -E ig run trace_tcpconnect:latest --verify-image=false --fields=src,dst
INFO[0000] Experimental features enabled
Error: pre-starting operator "cli": setting fields: column "src" is invalid
```

and the endpoints weren't present on the json output:

```bash
$ sudo -E ig run trace_tcpconnect:latest --verify-image=false -o json | jq
INFO[0000] Experimental features enabled
{
  "gid": 0,
  "k8s": {
    "containerName": "",
    "hostnetwork": false,
    "namespace": "",
    "node": "",
    "podName": ""
  },
  "latency": 0,
  "mntns_id": 4026534293,
  "pid": 137803,
  "runtime": {
    "containerId": "31fc2dd570cef84e2f9131e662833b6318cf5dac4d753c0854767e41419743c1",
    "containerImageDigest": "",
    "containerImageName": "wbitt/network-multitool",
    "containerName": "netem",
    "runtimeName": "docker"
  },
  "task": "wget",
  "timestamp": "2024-06-25T15:56:52.668658281-05:00",
  "timestamp_raw": 1719349012668658200,
  "uid": 0
}
```

#### After 

The columns output shows the "summary" version for dst and src columns by default, all other subfields are hidden:

(It's still to improve the annotations / attributes to avoid trimming these fields)

```bash 
$ sudo -E ig run trace_tcpconnect:latest --verify-image=false
INFO[0000] Experimental features enabled
RUNTIME.CONTAINERNAME   SRC          DST          TASK         PID          UID          GID          MNTNS_ID TIMESTAMP
netem                   172.17.0.2:4 1.1.1.1:80   wget         136059       0            0            40265342 2024-06-25T15:38:18.71826
netem                   172.17.0.2:5 1.1.1.1:443  wget         136059       0            0            40265342 2024-06-25T15:38:19.43726
netem                   172.17.0.2:3 1.1.1.1:443  wget         136059       0            0            40265342 2024-06-25T15:38:20.3098
```

```bash
$ sudo -E ig run trace_tcpconnect:latest --verify-image=false --fields=src,dst
INFO[0000] Experimental features enabled
SRC                                                                  DST
172.17.0.2:35584                                                     1.1.1.1:80
172.17.0.2:48386                                                     1.1.1.1:443
172.17.0.2:36782                                                     1.0.0.1:443
```

The json output now shows the fields in this way: (non endpoints fields were manually removed to reduce noise)

```bash 
$ sudo -E ig run trace_tcpconnect:latest --verify-image=false -o json | jq
INFO[0000] Experimental features enabled
{
  "dst": {
    "addr": "1.1.1.1",
    "port": 80,
    "proto": 6,
    "version": 4
  }, 
  "src": {
    "addr": "172.17.0.2",
    "port": 38566,
    "proto": 6,
    "version": 4
  },
}
```

Fixes #3034 
Fixes #2813
Fixes #2816

### TODO
- [ ] Provided human-friendly format for proto (future PR)