@mauriciovasquezbernal 2024-06-27
> Regarding the summary field: adding an annotation "columns.template: ipaddrport" could work.

It seems to be ignoring it. I'll investigate more later on (after this is merged)